requirements.txt
# Core scientific computing libraries
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Geospatial libraries
geopandas>=0.9.0
rasterio>=1.2.0
fiona>=1.8.0
shapely>=1.7.0
folium>=0.12.0

# Optimization libraries
pulp>=2.5.0
scipy>=1.7.0

# Database and data management
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
influxdb-client>=1.28.0
pandera>=0.9.0

# Web framework for dashboard
streamlit>=1.10.0
dash>=2.5.0
dash-bootstrap-components>=1.1.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.5.0

# Hydrological modeling
flopy>=3.3.5
pyemu>=1.1.0

# Configuration and utilities
pyyaml>=5.4.0
click>=8.0.0
tqdm>=4.62.0
loguru>=0.6.0
jsonschema>=3.2.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0